Change the algorithm
Wordle does not behave as the instructions indicate. If the word is
TRACE and you guess TRACT, the outcome is ggggb not ggggy

This change takes care of that. It also chagnes the datastructure for
storing the information so far. Instead of greens, yellows and blacks,
we now simply store the previous guesses and the outcomes, and then use
those to validate new guesses